requirements.txt
Django
python-dotenv
supabase
psycopg[binary]
whitenoise[brotli]
gunicorn
uvicorn
djangorestframework